Crypto functionality should now be working properly across ts and rs
