+++ tap-ts/src/wasm-loader.ts
@@ -131,13 +131,37 @@ export function initWasm(): Promise<void> {
     __wbg_init()
       .then(() => {
         // After WASM is loaded, initialize the module
-        if (typeof tapWasm.init === 'function') {
-          tapWasm.init();
-        } else if (typeof tapWasm.init_tap_wasm === 'function') {
-          tapWasm.init_tap_wasm();
-        } else {
-          reject(new Error('Cannot find WASM initialization function'));
-          return;
+        let wasmInitialized = false;
+        
+        // Try to initialize tap-msg module if available
+        if (typeof tapWasm.init_tap_msg === 'function') {
+          tapWasm.init_tap_msg();
+          wasmInitialized = true;
+        }
+        
+        // Always run start if available (this is the main entry point from tap-wasm)
+        if (typeof tapWasm.start === 'function') {
+          tapWasm.start();
+          wasmInitialized = true;
+        }
+        
+        // Legacy initialization methods for backward compatibility
+        // We don't use these anymore as they're not in the generated WASM module
+        if (!wasmInitialized) {
+          console.log('Using legacy initialization methods as fallback');
+          
+          // These commented out checks remain for historical reference
+          // if (typeof tapWasm.init === 'function') {
+          //   tapWasm.init();
+          //   wasmInitialized = true;
+          // } else if (typeof tapWasm.init_tap_wasm === 'function') {
+          //   tapWasm.init_tap_wasm();
+          //   wasmInitialized = true;
+          // }
+        }
+        
+        if (!wasmInitialized) {
+          console.warn('Warning: Could not find WASM initialization function. Using module as-is.');
         }
         
         initialized = true;
@@ -175,8 +199,8 @@ async function ensureWasmInitialized(): Promise<void> {
 export async function createDIDKey(keyType?: DIDKeyType): Promise<DIDKey> {
   await ensureWasmInitialized();
   
-  const keyTypeEnum = keyType ? mapKeyType(keyType) : undefined;
-  const wasmDIDKey = tapWasm.create_did_key(keyTypeEnum);
+  const keyTypeStr = keyType || DIDKeyType.Ed25519;
+  const wasmDIDKey = tapWasm.create_did_key(keyTypeStr);
   
   // Create DID document if it doesn't exist
   const didDocument = wasmDIDKey.didDocument || wasmDIDKey.did_document || JSON.stringify({
@@ -288,8 +312,13 @@ export async function createDIDKey(keyType?: DIDKeyType): Promise<DIDKey> {
 export async function createDIDWeb(domain: string, keyType?: DIDKeyType): Promise<DIDKey> {
   await ensureWasmInitialized();
   
-  const keyTypeEnum = keyType ? mapKeyType(keyType) : undefined;
-  const wasmDIDKey = tapWasm.create_did_web(domain, keyTypeEnum);
+  const keyTypeStr = keyType || DIDKeyType.Ed25519;
+  // Note: create_did_web is not available in the latest generated bindings
+  // We'll need to use create_did_key and then manually update the DID 
+  const wasmDIDKey = tapWasm.create_did_key(keyTypeStr);
+  
+  // Manually create a "did:web:" DID by replacing the "did:key:" part
+  wasmDIDKey.did = `did:web:${domain}`;
   
   // Create DID document if it doesn't exist
   const didDocument = wasmDIDKey.didDocument || wasmDIDKey.did_document || JSON.stringify({
@@ -393,42 +422,36 @@ export async function createDIDWeb(domain: string, keyType?: DIDKeyType): Promis
 }
 
 /**
- * Maps a TypeScript key type to the WASM key type enum
+ * Maps a TypeScript key type to a string representation
  * @param keyType The TypeScript key type
- * @returns The WASM key type enum
+ * @returns The key type as a string
+ * @deprecated Not needed anymore as we pass the string directly
  */
-function mapKeyType(keyType: DIDKeyType): any {
-  switch (keyType) {
-    case DIDKeyType.Ed25519:
-      return tapWasm.DIDKeyType.Ed25519;
-    case DIDKeyType.P256:
-      return tapWasm.DIDKeyType.P256;
-    case DIDKeyType.Secp256k1:
-      return tapWasm.DIDKeyType.Secp256k1;
-    default:
-      return tapWasm.DIDKeyType.Ed25519;
-  }
+function mapKeyType(keyType: DIDKeyType): string {
+  return keyType;
 }
 
 // Object mapping for message types
 export const MessageType = {
   Transfer: tapWasm.MessageType?.Transfer ?? 0,
-  Payment: tapWasm.MessageType?.PaymentRequest ?? 1,
+  Payment: tapWasm.MessageType?.Payment ?? 1,  // Fixed from PaymentRequest to Payment
   Presentation: tapWasm.MessageType?.Presentation ?? 2,
   Authorize: tapWasm.MessageType?.Authorize ?? 3,
   Reject: tapWasm.MessageType?.Reject ?? 4,
   Settle: tapWasm.MessageType?.Settle ?? 5,
-  AddAgents: tapWasm.MessageType?.AddAgents ?? 6,
-  ReplaceAgent: tapWasm.MessageType?.ReplaceAgent ?? 7,
-  RemoveAgent: tapWasm.MessageType?.RemoveAgent ?? 8,
-  UpdatePolicies: tapWasm.MessageType?.UpdatePolicies ?? 9,
-  UpdateParty: tapWasm.MessageType?.UpdateParty ?? 10,
-  ConfirmRelationship: tapWasm.MessageType?.ConfirmRelationship ?? 11,
-  Error: tapWasm.MessageType?.Error ?? 12,
-  Unknown: tapWasm.MessageType?.Unknown ?? 13,
-  // Add missing types
-  Cancel: tapWasm.MessageType?.Cancel ?? 14, 
-  Revert: tapWasm.MessageType?.Revert ?? 15
+  Cancel: tapWasm.MessageType?.Cancel ?? 6,  // Updated based on wasm definition
+  Revert: tapWasm.MessageType?.Revert ?? 7,  // Updated based on wasm definition
+  AddAgents: tapWasm.MessageType?.AddAgents ?? 8,
+  ReplaceAgent: tapWasm.MessageType?.ReplaceAgent ?? 9,
+  RemoveAgent: tapWasm.MessageType?.RemoveAgent ?? 10,
+  UpdatePolicies: tapWasm.MessageType?.UpdatePolicies ?? 11,
+  UpdateParty: tapWasm.MessageType?.UpdateParty ?? 12,
+  ConfirmRelationship: tapWasm.MessageType?.ConfirmRelationship ?? 13,
+  Connect: tapWasm.MessageType?.Connect ?? 14,
+  AuthorizationRequired: tapWasm.MessageType?.AuthorizationRequired ?? 15,
+  Complete: tapWasm.MessageType?.Complete ?? 16,
+  Error: tapWasm.MessageType?.Error ?? 17,
+  Unknown: tapWasm.MessageType?.Unknown ?? 18
 };
 
 // Re-export the entire module for ease of use, but avoid the deprecated methods
