+++ tap-wasm/Cargo.toml
@@ -28,7 +28,7 @@ web-sys = { version = "0.3.64", features = ["console"] }
 chrono = { version = "0.4", features = ["wasmbind"] }
 getrandom = { workspace = true, features = ["js"] }
 base64 = "0.21"
-ed25519-dalek = { version = "2.0.0", features = ["wasm-bindgen", "rand_core"] }
+ed25519-dalek = { version = "2.0.0", features = ["rand_core"] }
 
 [dependencies.wee_alloc]
 version = "0.4.5"
